[FIX]Pequeñas correcciones relacionadas con informes

+++ sale_volume/models/account_invoice_line.py
@@ -10,7 +10,7 @@ class AccountInvoiceLine(models.Model):
 
     _inherit = 'account.invoice.line'
 
-    product_uom_unit = fields.Float('units')
+    product_uom_unit = fields.Integer('units')
     escuadria = fields.Char('Escuadría')
     escuadria_float = fields.Float(compute='_compute_escuadria_float',
                                    store=True)
